+++ src/wrapper.cpp
@@ -191,6 +191,14 @@ List wrap_smallest_common_network(List address1_r, List address2_r) {
   return IpNetworkVector::smallestCommonNetwork(address1, address2).encodeR();
 }
 
+// [[Rcpp::export]]
+List wrap_summarize_address_range(List address1_r, List address2_r) {
+  IpAddressVector address1(address1_r);
+  IpAddressVector address2(address2_r);
+
+  return IpNetworkVector::summarizeAddressRange(address1, address2);
+}
+
 
 /*----------------------*
  *  Reserved addresses  *