+++ NEWS.md
@@ -2,9 +2,13 @@
 
 ## New features
 
-* New `common_network()` function to find the smallest network containing two addresses
-* `prefix_length()` can now accept an `ip_address()` vector of netmasks and/or hostmasks. This makes it possible to construct an `ip_network()` or `ip_interface()` from non-CIDR notation.
-  * E.g. `ip_network(ip_address("192.0.2.0"), prefix_length(ip_address("255.255.255.0")))`
+This release adds support for constructing `ip_network()` vectors without using CIDR notation.
+
+* New `common_network()` function finds the smallest network containing two addresses
+* New `summarize_address_range()` function lists the constituent networks of an address range
+* `prefix_length()` can now infer the prefix length from an `ip_address()` vector of netmasks and/or hostmasks. This makes it possible to construct an `ip_network()` vector like so:
+  * `ip_network(ip_address("192.0.2.0"), prefix_length(ip_address("255.255.255.0")))`
+  * `ip_network(ip_address("192.0.2.0"), prefix_length(ip_address("0.0.0.255")))`
 
 ## Minor improvements and fixes